terminal bootstrap
